requirements.txt
# requirements_server.txt
# SAMM-Anon 服务器环境依赖

# =============================================================================
# Core (版本锁定，确保兼容性)
# =============================================================================
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.3

# =============================================================================
# Audio Processing
# =============================================================================
librosa==0.10.1
soundfile==0.12.1
pyworld==0.3.4
resampy==0.4.2

# =============================================================================
# Machine Learning
# =============================================================================
scikit-learn==1.3.2
faiss-gpu==1.7.4       # GPU 版本，无 GPU 改为 faiss-cpu

# =============================================================================
# Data Storage & I/O
# =============================================================================
h5py==3.10.0
kaldiio==2.18.0
pandas==2.1.3

# =============================================================================
# Speech Processing
# =============================================================================
speechbrain==1.0.0

# =============================================================================
# Utilities
# =============================================================================
tqdm==4.66.1
pyyaml==6.0.1
tensorboard==2.15.1
matplotlib==3.8.2

# =============================================================================
# Evaluation (可选)
# =============================================================================
jiwer==3.0.3          # WER 计算

# =============================================================================
# Development (可选)
# =============================================================================
ipython>=8.0.0
jupyter>=1.0.0